refactor: 파일 분리및 타입 네이밍 변경
